Simplify controllers file structure
