+++ epa-avert-webservice/app/routes.js
@@ -1,8 +1,6 @@
 const route = require('koa-route');
 
-const rdf = require('./controllers/rdf');
-const eere = require('./controllers/eere');
-const displacement = require('./controllers/displacement');
+const controllers = require('./controllers');
 
 const routes = [
   route.get('/api/v1', (ctx) => {
@@ -10,26 +8,26 @@ const routes = [
   }),
 
   // debugging only (not called in web app)
-  route.get('/api/v1/rdf', rdf.list),
+  route.get('/api/v1/rdf', controllers.rdf.list),
 
   // web app method: fetchRegion()
   // (from panel 1, user clicks 'Set EE/RE Impacts' button)
-  route.get('/api/v1/rdf/:region', rdf.show),
+  route.get('/api/v1/rdf/:region', controllers.rdf.show),
 
   // debugging only (not called in web app)
-  route.get('/api/v1/eere', eere.list),
+  route.get('/api/v1/eere', controllers.eere.list),
 
   // web app method: fetchRegion()
   // (from panel 1, user clicks 'Set EE/RE Impacts' button)
-  route.get('/api/v1/eere/:region', eere.show),
+  route.get('/api/v1/eere/:region', controllers.eere.show),
 
   // web app method: calculateDisplacement()
   // (from panel 2, user clicks 'Get Results' button)
-  route.post('/api/v1/generation', displacement.calculateGeneration),
-  route.post('/api/v1/so2', displacement.calculateSo2),
-  route.post('/api/v1/nox', displacement.calculateNox),
-  route.post('/api/v1/co2', displacement.calculateCo2),
-  route.post('/api/v1/pm25', displacement.calculatePm25),
+  route.post('/api/v1/generation', controllers.displacement.calculateGeneration),
+  route.post('/api/v1/so2', controllers.displacement.calculateSo2),
+  route.post('/api/v1/nox', controllers.displacement.calculateNox),
+  route.post('/api/v1/co2', controllers.displacement.calculateCo2),
+  route.post('/api/v1/pm25', controllers.displacement.calculatePm25),
 ];
 
 module.exports = routes;
